requirements.txt
Django==2.2
django-ckeditor==6.0.0
django-countries==7.1
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-modeltranslation==0.17.1
django-money==1.3.1
django-phonenumber-field==5.1.0
phonenumbers==8.12.21
Pillow==8.2.0
psycopg2==2.8.6
py-moneyed==0.8.0
pytz==2021.1
pyuca==1.2
six==1.15.0
sqlparse==0.4.1
